requirements.txt
google-endpoints==2.4.4
google-endpoints-api-management==1.3.0
six==1.11.0
google-api-python-client==1.6.4
google-cloud==0.23.0
pytz==2016.10